Minor Cleanup Refactor
