+++ src/CofreeBot.hs
@@ -1,6 +1,8 @@
-module CofreeBot ( module CofreeBot.Session
+module CofreeBot ( module CofreeBot.Bot
+                 , module CofreeBot.Session
                  , module CofreeBot.Plugins
                  ) where
 
-import CofreeBot.Session
+import CofreeBot.Bot
 import CofreeBot.Plugins
+import CofreeBot.Session
